Rename proto_util to grpc_client
